jueves dda
ada

+++ ObligatorioDDA/src/Dominio/Fachada.java
@@ -38,7 +38,7 @@ public class Fachada {
         return sUsuario.LoginAdmin(nombreCompleto,password);
     }
 
-    public void agregarJugador(String cedula, String password, String nombre, int saldo) {
+    public void agregarJugador(String cedula, String password, String nombre, Double saldo) {
         sUsuario.agregarJugador(nombre,cedula,password,saldo);
        
     }
